+++ frontend_tests/node_tests/stream_list.js
@@ -20,6 +20,9 @@ set_global('message_store', {
 
 var stream_list = require('js/stream_list.js');
 
+// TODO: split out topic_list stuff to its own test module.
+var topic_list = require('js/topic_list.js');
+
 var jsdom = require("jsdom");
 var window = jsdom.jsdom().defaultView;
 global.$ = require('jquery')(window);
@@ -33,7 +36,7 @@ global.use_template('sidebar_private_message_list');
 global.use_template('stream_sidebar_row');
 global.use_template('stream_privacy');
 
-(function test_build_subject_list() {
+(function test_topic_list_build_list() {
     var stream = "devel";
     var active_topic = "testing";
     var max_topics = 5;
@@ -46,8 +49,8 @@ global.use_template('stream_privacy');
         return 1;
     };
 
-    var topic_html = stream_list._build_subject_list(stream, active_topic, max_topics);
-    global.write_test_output("test_build_subject_list", topic_html);
+    var topic_html = topic_list.build_list(stream, active_topic, max_topics);
+    global.write_test_output("test_topic_list_build_list", topic_html);
 
     var topic = $(topic_html).find('a').text().trim();
     assert.equal(topic, 'coding');
